+++ tests/PolicyTest.php
@@ -9,6 +9,7 @@ use Neat\Object\Test\Helper\GroupUser;
 use Neat\Object\Test\Helper\HardDelete;
 use Neat\Object\Test\Helper\NoEntity;
 use Neat\Object\Test\Helper\TimeStamps;
+use Neat\Object\Test\Helper\Type;
 use Neat\Object\Test\Helper\User;
 use PHPUnit\Framework\TestCase;
 use ReflectionProperty;
@@ -16,7 +17,16 @@ use RuntimeException;
 
 class PolicyTest extends TestCase
 {
-    /** @noinspection PhpDocMissingThrowsInspection */
+    /**
+     * Test factory method
+     */
+    public function testFactory()
+    {
+        $policy = new Policy();
+        $this->assertNull($policy->factory(User::class));
+        $this->assertSame([Type::class, 'createFromArray'], $policy->factory(Type::class));
+    }
+
     /**
      * Create property
      *